+++ db4e/Panes/XMRigRemotePane.py
@@ -13,7 +13,6 @@ from textual.containers import Container, Horizontal, Vertical, ScrollableContai
 from textual.widgets import Label, Select
 
 from db4e.Modules.XMRigRemote import XMRigRemote
-
 from db4e.Widgets.HashratePlot import HashratePlot
 
 from db4e.Constants.DLabel import DLabel
@@ -30,14 +29,14 @@ class XMRigRemotePane(Container):
     ip_addr_label = Label("", id=DForm.IP_ADDR_LABEL, classes=DForm.STATIC)
     hashrate_label = Label("", id=DForm.HASHRATE_LABEL, classes=DForm.STATIC)
     uptime_label = Label("", id=DForm.UPTIME_LABEL, classes=DForm.STATIC)
-    #hashrate_plot = HashratePlot(DLabel.HASHRATE, id=DField.HASHRATE_PLOT)
-    xmrig_remote = None
+    hashrate_plot = HashratePlot("Hashrates", id=DField.HASHRATE_PLOT)
+    select_widget = Select(compact=True, id=DForm.TIMES, options=DSelect.SELECT_LIST)
+    xmrig = None
 
 
     def compose(self):
         # Remote P2Pool daemon deployment form
-        INTRO = f"View information about the [cyan]{DLabel.XMRIG_REMOTE}[/] deployment here."
-
+        INTRO = f"View information about the [cyan]{DLabel.XMRIG_REMOTE}[/] deployment."
 
         yield Vertical(
             ScrollableContainer(
@@ -56,27 +55,37 @@ class XMRigRemotePane(Container):
                     Horizontal(
                         Label(DLabel.UPTIME, classes=DForm.FORM_LABEL_20),
                         self.uptime_label),
-                    classes=DForm.FORM_4, id=DForm.FORM_FIELD)),
+                    classes=DForm.FORM_4, id=DForm.FORM_FIELD),
+
+                Vertical(
+                    self.select_widget,
+                    classes=DForm.SELECT_BOX),
+
+                Vertical(
+                    self.hashrate_plot,
+                    classes=DForm.PANE_BOX)),
 
                 classes=DForm.PANE_BOX)
-        
+
+
     def on_select_changed(self, event: Select.Changed) -> None:
         selected_time = event.value
-        #hashrate_widget = self.query_one("#" + DField.HASHRATE_PLOT)
-        #hashrate_widget.update_time_range(selected_time)
+        self.hashrate_plot.update_time_range(selected_time)
 
 
     def set_data(self, xmrig: XMRigRemote):
-        #print(f"XMRig:set_data(): {xmrig}")
         self.xmrig = xmrig
         self.instance_label.update(xmrig.instance())
         self.ip_addr_label.update(xmrig.ip_addr())
         self.hashrate_label.update(str(xmrig.hashrate()) + " " + DLabel.H_PER_S)
         self.uptime_label.update(xmrig.uptime())
 
-        #hashrate_widget = self.query_one("#" + DField.HASHRATE_PLOT)
-        #hashrate_widget.load_all_data(xmrig.hashrates())
-        #hashrate_widget.update_time_range(self.selected_time)
+        data = xmrig.hashrates()
+        days = data[DField.DAYS]
+        hashrates = data[DField.VALUES]
 
-        #select_widget = self.query_one("#" + DForm.TIMES)
-        #select_widget.value = self.selected_time
+        plot = self.query_one("#" + DField.HASHRATE_PLOT, HashratePlot)
+        plot.load_data(days=days, hashrates=hashrates)
+        plot.hashrate_plot()
+
+